requirements.txt
firebase-admin==4.1.0
google-cloud-logging==1.15.0
google-auth==1.14.1
google-api-python-client==1.8.2
requests==2.23.0